Fix to Program input and File System
There was an issue where when loading a program, it would take the hex value of the number rather than the literal of it because it should be input in hex. So it would read 10 as ten, then convert that to 0A, instead of reading as 10. Also i fixed the file system. There was an issue with the TSB.